Move node/way/relation/old_* controllers into the api namespace
